+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed parent XML element to be Tumor instead of Patient for seerCaseSpecificCod (1914) and seerOtherCod (1915) in N18 dictionary. 
 - Added new methods on Patient and Tumor to remove a given data item.
+- Added option to standalone program to extract user-defined dictionaries into CSV files.
 - Updated embedded JRE to Java 12.0.1 (this only affects the standalone program).
 
 **Version 5.2**